+++ R/checkFormat.r
@@ -83,7 +83,7 @@ checkFormat.GADSdat <- function(GADSdat, type = "SPSS", changeFormat = TRUE) {
     })
   }
 
-  lablengths <- sapply(names(GADSdat$dat), function(gg) unique(na.omit(labels$format[labels$varName==gg])))
+  lablengths <- sapply(names(GADSdat$dat), function(gg) unique(stats::na.omit(labels$format[labels$varName==gg])))
 
   lablengths_r <- gsub("\\.0$", "", lablengths)
   for(uu in seq(along=lablengths)) {