+++ NEWS.md
@@ -1,9 +1,10 @@
 # eatGADS 0.17.0.9000
 
+* `checkFormat()` for checking and modifying`SPSS` format type compatability with actual data
 * `check4SPSS()` for checking `SPSS` meta data conventions and requirements
 * `changeVarNames()` and `applyChangeMeta()` now enforce `SQLite3` naming conventions
 * bug fixes `write_spss2()`
-* bug fix `applyChangeMeta()`: now recognices differences in `NAs` in old columns when checking the change table
+* bug fix `applyChangeMeta()`: now recognizes differences in `NAs` in old columns when checking the change table
 
 
 # eatGADS 0.17.0