+++ NEWS.md
@@ -1,6 +1,7 @@
 # eatGADS 0.17.0.9000
 
 * `applyChangeMeta()` and `changeValLabels()` now automatically assign `"valid"` to the `missings` columns for labeled values
+* bug fix `equalGADS()` (meta data differences solely due to different meta data sorting are now ignored) 
 * `equalGADS()` can now exclude certain meta data columns from the comparison
 * `checkFormat()` for checking and modifying`SPSS` format type compatability with actual data
 * `check4SPSS()` for checking `SPSS` meta data conventions and requirements